+++ src/mlrda.jl
@@ -72,7 +72,7 @@ function predict(object::Mlrda, X)
     m = size(X, 1)
     zp = predict(object.fm, X).pred
     z =  mapslices(argmax, zp; dims = 2) 
-    pred = reshape(replacebylev(z, object.lev), m, 1)
+    pred = reshape(replacebylev2(z, object.lev), m, 1)
     (pred = pred, posterior = zp)
 end